# src/backend/services/wireless/wireless-api-spec.yaml
openapi: 3.0.3
info:
  title: Wireless Service API
  description: |
    Multi-vendor wireless network management API supporting SSID management, 
    access point deployment, and security policy configuration for Cisco Meraki MR, 
    FortiAP, Cisco Catalyst WiFi, and Aruba access points.
  version: 1.0.0
  contact:
    name: Network Service Platform
    email: support@networkplatform.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.networkplatform.com/v1
    description: Production server
security:
  - bearerAuth: []
  - apiKey: []
paths:
...
  /wireless/security-profiles:
    get:
      summary: List security profiles
      description: Retrieve all configured security profiles
      tags:
        - Security Profiles
      responses:
        '200':
          description: List of security profiles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SecurityProfile'

    post:
      summary: Create security profile
      description: Create a new security profile template
      tags:
        - Security Profiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityProfileRequest'
      responses:
        '201':
          description: Security profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityProfile'
  /wireless/security-profiles/{profileId}:
    get:
      summary: Get security profile
      tags:
        - Security Profiles
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Security profile details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityProfile'

    put:
      summary: Update security profile
      tags:
        - Security Profiles
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSecurityProfileRequest'
      responses:
        '200':
          description: Security profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityProfile'

  # VLAN Integration
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    SecurityProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        authType:
          type: string
          enum: [open, wep, wpa2_psk, wpa2_enterprise, wpa3_psk, wpa3_enterprise]
        encryption:
          type: string
          enum: [none, wep, tkip, aes, mixed]
        passphrase:
          type: string
          format: password
        radiusConfig:
          type: object
          properties:
            primaryServer:
              type: object
              properties:
                host:
                  type: string
                port:
                  type: integer
                  default: 1812
                secret:
                  type: string
                  format: password
            secondaryServer:
              type: object
              properties:
                host:
                  type: string
                port:
                  type: integer
                  default: 1812
                secret:
                  type: string
                  format: password
        certificateConfig:
          type: object
          properties:
            caCertificate:
              type: string
            serverCertificate:
              type: string
            privateKey:
              type: string
              format: password
        vendorSpecific:
          type: object
        isTemplate:
          type: boolean
          default: false
        metadata:
          type: object
          properties:
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            createdBy:
              type: string
            description:
              type: string
    CreateSecurityProfileRequest:
      type: object
      required:
        - name
        - authType
      properties:
        name:
          type: string
          maxLength: 64
        authType:
          type: string
          enum: [open, wep, wpa2_psk, wpa2_enterprise, wpa3_psk, wpa3_enterprise]
        encryption:
          type: string
          enum: [none, wep, tkip, aes, mixed]
        passphrase:
          type: string
          format: password
          minLength: 8
          maxLength: 63
        radiusConfig:
          type: object
          properties:
            primaryServer:
              type: object
              required: [host, secret]
              properties:
                host:
                  type: string
                port:
                  type: integer
                  default: 1812
                secret:
                  type: string
                  format: password
            secondaryServer:
              type: object
              properties:
                host:
                  type: string
                port:
                  type: integer
                  default: 1812
                secret:
                  type: string
                  format: password
        certificateConfig:
          type: object
        vendorSpecific:
          type: object
        isTemplate:
          type: boolean
          default: false
        description:
          type: string
          maxLength: 255
    UpdateSecurityProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 64
        authType:
          type: string
          enum: [open, wep, wpa2_psk, wpa2_enterprise, wpa3_psk, wpa3_enterprise]
        encryption:
          type: string
          enum: [none, wep, tkip, aes, mixed]
        passphrase:
          type: string
          format: password
          minLength: 8
          maxLength: 63
        radiusConfig:
          type: object
        certificateConfig:
          type: object
        vendorSpecific:
          type: object
        isTemplate:
          type: boolean
        description:
          type: string
          maxLength: 255
...